Fix add url to batch of samples

+++ src/main/java/com/ttps/laboratorio/dto/request/UrlDTO.java
@@ -2,12 +2,15 @@ package com.ttps.laboratorio.dto.request;
 
 import java.util.ArrayList;
 import java.util.List;
+
 import javax.validation.constraints.NotBlank;
+
+import org.springframework.validation.annotation.Validated;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.validation.annotation.Validated;
 
 @Validated
 @Getter
@@ -19,6 +22,6 @@ public class UrlDTO {
 	@NotBlank(message = "Url is required.")
 	private String url;
 
-	private List<Integer> failedSamples = new ArrayList<>();
+	private List<Long> failedSamples = new ArrayList<>();
 
 }
